Moved specs from components
