+++ app/assets/javascripts/components/IncidentCard.fixture.js
